backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hoang Le"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
langserve = {extras = ["all"], version = "^0.0.51"}
httpx = "^0.27.0"
pydantic = "^2.9.2"
langchain-core = "^0.3.17"
langchain-openai = "^0.2.8"
langgraph = "^0.2.46"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
